float max number from 3 values

+++ MaxNumber.cs
@@ -8,15 +8,15 @@ namespace GenericMaxNumber
 {
     public class MaxNumber
     {
-        public int first, second, third;
+        public float first, second, third;
 
-        public MaxNumber(int first, int second, int third)
+        public MaxNumber(float first, float second, float third)
         {
             this.first = first;
             this.second = second;
             this.third = third;
         }
-        public int GetMaxNum()
+        public float GetMaxNum()
         {
             if (first.CompareTo(second) == 1 && first.CompareTo(third) ==1)
                 return first;
